la perfezione dei barboni

+++ matron/src/hardware/screen/ssd1322.h
@@ -95,7 +95,7 @@ void ssd1322_deinit();
 void ssd1322_refresh();
 void ssd1322_update(cairo_surface_t * surface, bool should_translate_color);
 void ssd1322_set_brightness(uint8_t b);
-void ssd1322_set_contrast(uint8_t c);
+void ssd1322_set_contrast(int c);
 void ssd1322_set_display_mode(ssd1322_display_mode_t);
 void ssd1322_set_gamma(double g);
 void ssd1322_set_refresh_rate(uint8_t hz);
